requirements.txt
streamlit>=1.30.0
anthropic>=0.18.0
google-generativeai>=0.3.0
openai>=1.12.0
python-dotenv>=1.0.0
reportlab>=4.0.0
aiohttp>=3.9.0
pydantic>=2.5.0
sqlalchemy>=2.0.0

# Text-to-Speech dependencies
google-cloud-texttospeech>=2.14.0
gtts>=2.3.0
